Added tools for CI/CD integration
Added staging_build script which allows to build a specific
branch and component inside of the Cloud:Images:Factory:Staging
project. The script is expected to be called as follows:

    staging_build branch id workflow

branch is expected to be the pull request branch, id is a unique
identifier and expected to be the pull request number and workflow
is the name of a method that staging_build has to implement to
run the correct workflow associated with the pull request. Typically
you want to run all workflows and for now the image description
has one target which is cloud and therefore the script currently
implements one workflow named component_cloud. This allows to run
the staging build like shown in the following example:

    staging_build factory 0815 component_cloud

In a github actions environment the above call could be part of
an actions setup. As we currently haven't decided for a CI/CD
system this PR only adds the staging implementation which still
needs to be called manually.